refactor: use manganato instead of mangakakalot #102
