+++ provider/generic/pages.go
@@ -1,7 +1,7 @@
 package generic
 
 import (
-	"github.com/gocolly/colly"
+	"github.com/gocolly/colly/v2"
 	"github.com/metafates/mangal/source"
 	"net/http"
 )
