+++ provider/provider.go
@@ -5,7 +5,7 @@ import (
 	"github.com/metafates/mangal/provider/custom"
 	"github.com/metafates/mangal/provider/generic"
 	"github.com/metafates/mangal/provider/mangadex"
-	"github.com/metafates/mangal/provider/mangakakalot"
+	"github.com/metafates/mangal/provider/manganato"
 	"github.com/metafates/mangal/provider/manganelo"
 	"github.com/metafates/mangal/provider/mangapill"
 	"github.com/metafates/mangal/source"
@@ -19,7 +19,7 @@ import (
 func init() {
 	for _, conf := range []*generic.Configuration{
 		manganelo.Config,
-		mangakakalot.Config,
+		manganato.Config,
 		mangapill.Config,
 	} {
 		conf := conf
